main/src/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pandas
numpy
faiss-cpu
gunicorn
